src/server/requirements.txt
alembic~=0.8.7
pyyaml~=3.11
sqlalchemy~=1.1.0b3
tornado~=4.4.1

pudb
